tests/wpt/tests/css/css-conditional/container-queries/top-layer-dialog.html

<!doctype html>
<title>CSS Container Queries Test: @container with modal dialog child</title>
<link rel="help" href="https://drafts.csswg.org/css-conditional-5/#container-queries">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/cq-testcommon.js"></script>
<style>
  #container {
    container-type: inline-size;
  }
  dialog {
    color: red;
  }
  @container (max-width: 200px) {
    dialog { color: green; }
  }
  @container (max-width: 100px) {
    dialog { color: lime; }
  }
</style>
<div id="container">
  <dialog id="dialog"></dialog>
</div>
<script>
  setup(() => assert_implements_size_container_queries());

  test(() => {
    assert_equals(getComputedStyle(dialog).color, "rgb(255, 0, 0)");
  }, "#container initially wider than 200px");

  test(() => {
    container.style.width = "200px";
    assert_equals(getComputedStyle(dialog).color, "rgb(0, 128, 0)");
  }, "#container changed to 200px");

  test(() => {
    dialog.showModal();
    assert_equals(getComputedStyle(dialog).color, "rgb(0, 128, 0)");
  }, "Modal dialog still has parent as query container while in top layer");

  test(() => {
    container.style.width = "100px";
    assert_equals(getComputedStyle(dialog).color, "rgb(0, 255, 0)");
  }, "Container changes width while dialog is in top layer");
</script>
